+++ nucleotides/util.py
@@ -1,3 +1,4 @@
+from __future__ import print_function
 import os, os.path
 
 import nucleotides.log      as log
@@ -22,7 +23,7 @@ def parse(doc, argv, opts = False):
 def create_application_state(task):
     path = os.path.join("nucleotides", task)
     bbx_util.mkdir_p(path)
-    return {'api'    : os.environ["NUCLEOTIDES_API"],
+    return {'api'    : get_environment_variable("NUCLEOTIDES_API"),
             'logger' : log.create_logger(os.path.join(path, "benchmark.log")),
             'path'   : path}
 
@@ -52,3 +53,11 @@ def sha_digest(filename):
         for chunk in iter(lambda: f.read(sha.block_size), b''):
             sha.update(chunk)
     return sha.hexdigest()
+
+def get_environment_variable(name):
+    import sys
+    if not name in os.environ:
+        print("Missing environment variable: {}".format(name), file=sys.stderr)
+        exit(1)
+    return os.environ[name]
+